portfolio-chatbot/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.35.0
openai==1.93.0
sentence-transformers==5.0.0
beautifulsoup4==4.13.4
requests==2.32.4
python-dotenv==1.1.1
pydantic==2.11.7
numpy==2.3.1
faiss-cpu==1.11.0
scikit-learn==1.6.1
transformers==4.47.0
torch==2.7.1
tiktoken==0.8.0
python-multipart==0.0.20
